% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{better_pheatmap}
\alias{better_pheatmap}
\title{Better Pheatmaps}
\usage{
better_pheatmap(ph)
}
\arguments{
\item{ph}{a pheatmap object}
}
\value{
a modified pheatmap object with lines filled with fill
colours.
}
\description{
Remove the greylines that plagues these plots by
modifying the raw ggplot grob.
}
