% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{calculate_cluster_corr}
\alias{calculate_cluster_corr}
\title{Calculate Cluster Correlation}
\usage{
calculate_cluster_corr(clust_assign, scale_mat, out_dir, prefix_str)
}
\arguments{
\item{clust_assign}{A two-column table of 'gene' and 'cluster'.}

\item{scale_mat}{A matrix with genes as rows and samples as
columns.}

\item{out_dir}{A string denoting the output directory to store
plots and tables.}

\item{prefix_str}{A string prefix for the filename}
}
\value{
A three-column table of 'gene', 'cluster' and 'score'.
}
\description{
Calculate mean expression of each sample for each
cluster to have gene centroids, and use these to calculate
correlation scores for each gene against the centroid.
}
\examples{
n <- 100
scale_mat <- matrix(rnorm(n**2, mean = 0), nrow = n)
rownames(scale_mat) <- paste0("G", 1:n)
colnames(scale_mat) <- paste0("S", 1:n)
clust_assign <- data.frame(
    gene = rownames(scale_mat),
    cluster = c(rep(1, n / 50), rep(2, n / 50)),
    value = rnorm(n, 100, 10)
)
ca <- calculate_cluster_corr(clust_assign, scale_mat, tempdir(), "red")
}
