% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{calculate_cluster_corr_i}
\alias{calculate_cluster_corr_i}
\title{Calculate Cluster Correlation for a single cluster}
\usage{
calculate_cluster_corr_i(clust_assign, scale_mat, i)
}
\arguments{
\item{clust_assign}{A two-column table of 'gene' and 'cluster'.}

\item{scale_mat}{A matrix with genes as rows and samples as
columns.}

\item{i}{A positive integer representing a cluster number.}
}
\value{
A three-column table of 'gene', 'cluster' and 'score'.
}
\description{
Calculate cluster correlation scores for a single
cluster of genes
}
\examples{
n <- 100
scale_mat <- matrix(rnorm(n**2, mean = 0), nrow = n)
rownames(scale_mat) <- paste0("G", 1:n)
colnames(scale_mat) <- paste0("S", 1:n)
clust_assign <- data.frame(
    gene = rownames(scale_mat),
    cluster = c(rep(1, n / 50), rep(2, n / 50)),
    value = rnorm(n, 100, 10)
)
cac <- calculate_cluster_corr_i(clust_assign, scale_mat, 2)
}
