% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{cluster_assignments}
\alias{cluster_assignments}
\title{Cluster Assignments}
\usage{
cluster_assignments(hm_now_drawn, matobj)
}
\arguments{
\item{hm_now_drawn}{A ComplexHeatmap object invoked with
\code{draw}.}

\item{matobj}{A matrix with genes as rownames.}
}
\value{
A two column data frame of Gene and Cluster in ascending
order of cluster number.
}
\description{
Get the cluster assignment for all genes
}
