% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{cluster_gene_plots}
\alias{cluster_gene_plots}
\title{Cluster Gene Plots}
\usage{
cluster_gene_plots(tab, score_thresh = 0, out_dir = "gene_cluster")
}
\arguments{
\item{tab}{A dataframe with long plotting data containing at least
the columns \code{cluster}, \code{condition}, \code{value},
\code{time}, and \code{score}. It can take normalised or scaled
values as input.}

\item{score_thresh}{numeric threshold to filter out genes with
correlation scores not matching these values}

\item{out_dir}{a string depicting the directory of where the plots
are deposited, with the prefix "gene_plots-k".}
}
\value{
ggplot object ribbon and line plots facetted by cluster.
}
\description{
Generate ribbon and line plots of gene expression data
split by cluster.
}
