% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{do_kmeans}
\alias{do_kmeans}
\title{Perform K-means Pairwise and Global}
\usage{
do_kmeans(
    norm_counts,
    transformed_counts,
    phenotype_data,
    genes,
    sample_columns,
    dsqres,
    k,
    out_dir
)
}
\arguments{
\item{norm_counts}{a matrix of normalized counts}

\item{transformed_counts}{a matrix of transformed counts, VST or
other produced by DESeq2}

\item{phenotype_data}{a data.frame of phenotype data which has the
same rows as the number of samples.}

\item{genes}{A list with at least 4 components: \code{cluster}, a vector
of genes to use for clustering, \code{highlight}, a vector of genes
to highlight in the heatmap; \code{interest}, a list of genes to
plot, where if NULL, use \code{highlight}, and if FALSE do not use;
\code{score_thresh} a vector of numerics depicting cluster score
thresholds to filter for high quality genes in each cluster
before plotting. Default is \code{c(0, 0.5, 0.9, 0.99)}}

\item{sample_columns}{a vector of sample names. If NULL (the
default), then use all samples}

\item{dsqres}{the output of the DESeq2 function \code{results},
usually called after performing a contrast.}

\item{k}{A positive integer to do k-means.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}
}
\value{
None. Produces only tables and plots in the output
directory.
}
\description{
Produce two K-means plots where one contains pairwise
comparisons between samples of interest, and one contains the
same pairwise comparisons but extended over all samples.
}
