% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{do_volcanos}
\alias{do_volcanos}
\title{Do volcano plots}
\usage{
do_volcanos(
    dsqres,
    top_genes_tohighlight,
    plot_title,
    outdir,
    volcano_params = NULL
)
}
\arguments{
\item{dsqres}{A DESeq2 Result object}

\item{top_genes_tohighlight}{A vector of genes to highlight in the
volcano plots}

\item{plot_title}{A string depicting the title of the gplots}

\item{outdir}{An output directory path fro the plots}

\item{volcano_params}{A list of two components: "global" and
"zoomed" each with "curve" and "curve_show" components
describing the curve to fit to the volcano plots by \code{sd}
(standard deviation from center mean), \code{sc} vertical scale
factor, \code{offset} vertical offset. Extra params for the
\code{zoomed} component are \code{ylim}, \code{padj} an upper
limit on the adjusted p-value, \code{lfc} and an upper-limit on
the log2 fold change. If \code{curve_show} is set to FALSE in
either \code{global} or \code{zoomed} then no curve is shown.}
}
\value{
None.
}
\description{
Generate General and Zoomed volcano plots
}
\examples{
n <- 100
dsqres <- data.frame(
    gene = paste0("G", 1:n),
    data = rnorm(n, 5, 2),
    log2FoldChange = 1e-5,
    mLog10Padj = rnorm(n, 1000, 2),
    padj = 1/rnorm(n, 1000, 2)
)
top_genes_tohighlight <- paste0("G", (n-50):(n-20))
res <- RNASeqHelper:::do_volcanos(dsqres, top_genes_tohighlight,
                                "my plot", tempdir())
}
