% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Plots heatmap for n genes most differentially expressed in comparison specified by contrast, 
specific strains can be chosen in groups}
\usage{
draw_heatmap(data_edger, comparison, groups, n = 50)
}
\arguments{
\item{data_edger}{experiment data from edgeR analysis}

\item{comparison}{list containing table with pvalues for gene ordering}

\item{groups}{specific strains to show on heatmap}

\item{n}{number of genes shown on heatmap}
}
\value{
heatmap
}
\description{
Plots heatmap for n genes most differentially expressed in comparison specified by contrast, 
specific strains can be chosen in groups
}
