% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{gene_clusters_by_score}
\alias{gene_clusters_by_score}
\title{Gene clusters by score}
\usage{
gene_clusters_by_score(
    tab,
    score_thresh = c(0, 0.5, 0.9, 0.99),
    out_dir = "gene_cluster"
)
}
\arguments{
\item{tab}{A dataframe with long plotting data containing at least
the columns \code{gene}, \code{cluster}, \code{condition},
\code{value}, \code{time}, and \code{score}. It can take
normalised or scaled values as input.}

\item{score_thresh}{Vector of numerics depicting cluster score
thresholds to filter for high quality genes in each cluster
before plotting. Default is \code{c(0, 0.5, 0.9, 0.99)}}

\item{out_dir}{String representing the directory to store gene
plots. Default value is "gene_cluster"}
}
\value{
A single patchwork plot containing cluster gene plots for
different score thresholds. This plot is also saved to the
\code{out_dir} directory under the pattern
"gene_plots-k*_montage.svg"
}
\description{
Generate cluster gene plots for various score
thresholds
}
\examples{
n <- 100
tab <- data.frame(
    gene = paste0("G", 1:n),
    cluster = as.integer(rnorm(n, 5, 2) + 1),
    condition = c(rep("red", n/2), rep("green", n/2)),
    value = rnorm(n, 10, 2) + 1,
    time = as.integer(rnorm(n, 2, 0.5) + 1) * 5,
    score = as.integer(rnorm(n, 2, 0.5) + 1) / 4
)
plot <- gene_clusters_by_score(tab, c(0.2, 0.7), out_dir = tempdir())
}
