% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{gene_plots_by_gene}
\alias{gene_plots_by_gene}
\title{Gene plots by gene}
\usage{
gene_plots_by_gene(
    norm_long,
    scale_long,
    genes_of_interest,
    outprefix = "gene.lists",
    out_dir = "gene_lists"
)
}
\arguments{
\item{norm_long}{A long dataframe of normalized values. Must
contain columns of: gene, condition, time, replicate, value.}

\item{scale_long}{A long dataframe of scaled values. Same format as
\code{norm_long}.}

\item{genes_of_interest}{A named list of gene vectors. A list of
genes which are referenced by specific names. Unique plots will
be generated for list. e.g. list(mygeneset1 = c("Msgn1",
"Osr1", "Rspo3", "Fgf8", "Wnt3a"), mygeneset2 = c("Mesp1",
"Foxa2", "Sox17", "Lhx1", "Cer1"))}

\item{outprefix}{A string depicting the output prefix for the
plots. These will be appended with the title of each gene list
and for the normalized and scaled.}

\item{out_dir}{A string denoting the output directory to store
plots. Default is is "gene_lists".}
}
\value{
None.
}
\description{
Generate plots for genes of interest showing their
normalized and z-score scaled expression.
}
