% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_transcripts_from_gb}
\alias{get_transcripts_from_gb}
\title{will produce a data.frame with gene names, protein id and function and a note from genbank file}
\usage{
get_transcripts_from_gb(genbank, add_names = TRUE, uniprot_file)
}
\arguments{
\item{genbank}{genbank file read in using genbankr package}

\item{add_names}{add names from uniprot file? default TRUE}

\item{uniprot_file}{uniprot file containing protein name etc.}
}
\value{
data frame containing gene id, locus_tag, product, protein_id and note from genbank file
}
\description{
will produce a data.frame with gene names, protein id and function and a note from genbank file
}
