% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{heatmap_with_geneplots}
\alias{heatmap_with_geneplots}
\title{Heatmap with Gene plots}
\usage{
heatmap_with_geneplots(
    norm_counts,
    pheno_data,
    k,
    genes,
    out_dir = "heatmaps_k",
    heatprefix = "heatmap",
    prefix_title = "",
    width_in = 6,
    height_in = 6
)
}
\arguments{
\item{norm_counts}{A matrix containing normalised values.}

\item{pheno_data}{A data.frame of phenotype data.}

\item{k}{An integer for the k-value for kmeans.}

\item{genes}{A list with at least 3 components: \code{highlight}, a
vector of genes to highlight in the heatmap; \code{interest}, a
list of genes to plot, where if NULL, use \code{highlight}, and
if FALSE do not use; \code{score_thresh} a vector of numerics
depicting cluster score thresholds to filter for high quality
genes in each cluster before plotting. Default is \code{c(0,
    0.5, 0.9, 0.99)}}

\item{out_dir}{A character sequence depicting the directory}

\item{heatprefix}{A character sequence depicting the prefix for
heatmaps}

\item{prefix_title}{A string to prefix the title of heatmap.}

\item{width_in}{A positive integer for the number of inches in the
plot width.}

\item{height_in}{A positive integer for the number of inches in the
plot height.}
}
\value{
A list of two components; clustered tables, and scaled
matrix.
}
\description{
Generate a clustered heatmaps for a specific k means
value.
}
\examples{
options(bitmapType='cairo') ## disable plotting to X11
n <- 100
norm_counts <- matrix(rnorm(n**2, mean = 5), nrow = n)
pheno <- data.frame(sample=paste0("S", 1:n),
                    condition = c(rep("red", n / 2), rep("green", n / 2)),
                    time = as.integer(rnorm(n, 2, 0.5) + 1) * 5)
rownames(norm_counts) <- paste0("G", 1:n)
colnames(norm_counts) <- paste0("S", 1:n)
genes <- list(highlight = paste0("G", 1:10),
            interest = paste0("G", c(2,3)),
            score_thresh = c(0.2, 0.3))
res <- heatmap_with_geneplots(norm_counts, pheno, 2,
                            genes, out_dir = tempdir())
}
