% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{high_quality_genes}
\alias{high_quality_genes}
\title{Get a high quality genes list}
\usage{
high_quality_genes(sam_mat, min_occur = 3, min_detect = 10, out_dir)
}
\arguments{
\item{sam_mat}{A numeric matrix. Genes are rows and Samples as
columns. Typically this is the average sample matrix, with
replicate values summarized into mean expression.}

\item{min_occur}{A positive integer. A threshold (greater than or
equal to) the number of times a gene is detected in a sample,
across all samples (e.g. Gene X is only kept if it appears in
10 samples with a value greater than the \code{min_detect}
threshold. Default value is 3.}

\item{min_detect}{A positive integer. A threshold (greater than or
equal to) for the minimum expression a gene can have for it to
be a valid occurrence in a sample. Default values is 10.}

\item{out_dir}{A string depicting the output directory to place
tables.}
}
\value{
A vector of strings depicting a list of high quality genes.
A table of rowSums is written to a file matching
"smallestGroup*-detected*-keep_genes".
}
\description{
Filter a wide-format matrix for high quality genes
(rows) based on detectability thresholds and smallest
occurrence.
}
\examples{
sam_mat <- matrix(1:100, nrow = 10, dimnames = list(
    paste0("G", 1:10),
    paste0("C", 1:10)
))
keep <- high_quality_genes(sam_mat, 10, 7, "/tmp")
names(keep) == paste0("G", 7:10)
}
