% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{kmeans_heatmaps}
\alias{kmeans_heatmaps}
\title{Nice K-means Heatmap}
\usage{
kmeans_heatmaps(
    norms,
    trans,
    pheno,
    genes,
    dsqres,
    k,
    out_dir,
    heatprefix,
    plot_title,
    sample_columns = NULL
)
}
\arguments{
\item{norms}{a matrix of normalized counts}

\item{trans}{a matrix of transformed counts, VST or other produced
by DESeq2.}

\item{pheno}{a data.frame of phenotype data}

\item{genes}{A list with at least 4 components: \code{cluster}, a vector
of genes to use for clustering, \code{highlight}, a vector of genes
to highlight in the heatmap; \code{interest}, a list of genes to
plot, where if NULL, use \code{highlight}, and if FALSE do not use;
\code{score_thresh} a vector of numerics depicting cluster score
thresholds to filter for high quality genes in each cluster
before plotting. Default is \code{c(0, 0.5, 0.9, 0.99)}}

\item{dsqres}{the output of the DESeq2 function \code{results},
usually called after performing a contrast.}

\item{k}{A positive integer to do k-means.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}

\item{heatprefix}{String to prefix heatmap plot filenames}

\item{plot_title}{String depicting title to embed into plot}

\item{sample_columns}{a vector of sample names. If NULL (the
default), then use all samples}
}
\value{
None.
}
\description{
Produce a nice K-means clustered heatmap for both
normalized and transformed (DESeq2 VST corrected) matrices
}
