% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{num_clusters}
\alias{num_clusters}
\title{Number of clusters in tabs}
\usage{
num_clusters(tab)
}
\arguments{
\item{tab}{Dataframe. Must contain a \code{cluster} vector}
}
\value{
Positive integer.
}
\description{
Count the number of unique values in the
\code{cluster} column of a table
}
\examples{
RNASeqHelper:::num_clusters(
    data.frame(cluster = c(1, 1, 2, 3, 2))
)
}
