% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{pca_and_matrices}
\alias{pca_and_matrices}
\title{Print and Store PCA and Matrices}
\usage{
pca_and_matrices(res, out_dir = "deseq2")
}
\arguments{
\item{res}{Output of \code{run_deseq}.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}
}
\value{
None. Plots and tables and placed into output directory.
}
\description{
Takes the output of a DESeq2 run and generates
normalization tables and PCAs
}
