% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{read_antismash}
\alias{read_antismash}
\title{function reads from antismash files query genes with best subject matching genes}
\usage{
read_antismash(file)
}
\arguments{
\item{file}{antismash result file in txt}
}
\value{
data frame with antismash results with gene information and blast data 
(% identity, blast score evalue) for each found similar gene
}
\description{
function reads from antismash files query genes with best subject matching genes
}
