% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{read_cluster_file}
\alias{read_cluster_file}
\title{Reads cluster file prepared using clust program}
\usage{
read_cluster_file(file, genes_list)
}
\arguments{
\item{file}{clust result file}

\item{genes_list}{list of genes}
}
\value{
data frame with all genes with assigned cluster or NA
}
\description{
Reads cluster file prepared using clust program
}
