% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{rnaseqhelper}
\alias{rnaseqhelper}
\title{Run a full RNASeqhelper analysis}
\usage{
rnaseqhelper(
    tab,
    phenotype_data,
    out_dir = tempdir(),
    numer,
    denom,
    keep_params = list(),
    heat_params = list(),
    volcano_params = list()
)
}
\arguments{
\item{tab}{a matrix of samples (columns) and genes (rows)}

\item{phenotype_data}{a table with samples (rows) with extra
columns for annotation groups. One of these groups must be
condition and will be used to normalize the data.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}

\item{numer}{A string value to select sample columns from the
\code{condition} column that will be part of the numerator part
of the contrast.}

\item{denom}{A string value to select sample columns from the
\code{condition} column, that will be the denominator part of
the contrast.}

\item{keep_params}{A list of two components \code{min_occur} and
\code{min_detect} as defined by the default
\code{high_quality_genes} function.}

\item{heat_params}{A list of parameters to override the default
\code{pairwise_heatmap_volcano} function. If \code{NULL}, then
use the default.}

\item{volcano_params}{A list of parameters to override the default
\code{do_volcanos} function. If \code{NULL}, then use the
default.}
}
\value{
None.
}
\description{
Single function to call all other functions in the
analysis and perform a full analysis from a few starting
parameters
}
\examples{
n <- 1000
tab <- matrix(as.integer(rnorm(n**2, 1000, 500)), ncol = n)
tab <- tab - min(tab)
rownames(tab) <- paste0("G", 1:n)
colnames(tab) <- paste0("S", 1:n)
phenotype_data <- data.frame(
    sample = paste0("S", 1:n),
    condition = c(rep("red", n / 2), rep("green", n / 2)),
    time = as.integer(rnorm(n, 2, 0.5) + 1) * 5
)
rnaseqhelper(tab, phenotype_data,
    out_dir = tempdir(), "green", "red",
    heat_params = list(
        score_thresh = c(0.2, 0.5),
        kmeans_list = 2
    )
)
}
