% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{run_deseq}
\alias{run_deseq}
\title{Run DESeq with sensible defaults}
\usage{
run_deseq(tab, keep_genes, phenotype_data, out_dir)
}
\arguments{
\item{tab}{a matrix of samples (columns) and informative genes
(rows), ideally subset using the function
\code{high_quality_genes}}

\item{keep_genes}{a vector of genes to subset.}

\item{phenotype_data}{a table with samples (rows) with extra
columns for annotation groups. One of these groups must be
condition and will be used to normalize the data.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}
}
\value{
list of tables
}
\description{
Runs DESeq with filtering thresholds and shows PCA for
variance-stabilize-transformed data.
}
\examples{
ngene <- 1000
nsample <- 10
tab <- matrix(as.integer(rnorm(ngene * nsample, 1000, 500)), ncol = nsample)
tab <- tab - min(tab)
rownames(tab) <- paste0("G", seq(ngene))
colnames(tab) <- paste0("S", seq(nsample))
phenotype_data <- data.frame(
    sample = colnames(tab),
    condition = c(rep("red", nsample / 2), rep("green", nsample / 2)),
    time = as.integer(rnorm(nsample, 2, 0.5) + 1) * 5
)
keep_genes <- rownames(tab)
res <- run_deseq(tab, keep_genes, phenotype_data, tempdir())
}
