% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{single_gene_plot}
\alias{single_gene_plot}
\title{Single Gene Plot}
\usage{
single_gene_plot(
    long_data,
    genes_found,
    glist_name,
    ylab_text,
    title_text,
    out_dir,
    outprefix,
    filesuffix,
    scaley10
)
}
\arguments{
\item{long_data}{A tibble containing columns: gene, time, condition, value}

\item{genes_found}{A vector of gene names}

\item{glist_name}{A string for the group name for said vector of genes}

\item{ylab_text}{A string for Y-axis label text}

\item{title_text}{A string for the title text}

\item{out_dir}{A string denoting the output directory to store
plots. Default is is "gene_lists".}

\item{outprefix}{A string for the filename prefix}

\item{filesuffix}{A string depicting the filename suffix and extension.}

\item{scaley10}{A boolean on whether to scale the Y-axis by log10}
}
\value{
A ggplot2 object
}
\description{
Perform a plot of all genes for a single instance of
long table data and save to file, as well as return the plot.
}
