% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{single_kmeans_heatmap}
\alias{single_kmeans_heatmap}
\title{A single K-means Heatmap}
\usage{
single_kmeans_heatmap(
    scale_mat,
    k,
    top_genes,
    prefix_title,
    top_title,
    out_dir,
    heatprefix,
    width_in,
    height_in
)
}
\arguments{
\item{scale_mat}{A matrix of scaled values.}

\item{k}{A positive integer for the k value of kmeans}

\item{top_genes}{A list of genes to plot}

\item{prefix_title}{A string to prefix the title of heatmap.}

\item{top_title}{A character string for the title of the heatmap.}

\item{out_dir}{A character sequence depicting the directory}

\item{heatprefix}{A character sequence depicting the prefix for
heatmaps}

\item{width_in}{A positive integer for the number of inches in the
plot width.}

\item{height_in}{A positive integer for the number of inches in the
plot height.}
}
\value{
A list of two components; clustered tables, and scaled
matrix.
}
\description{
Plot a clustered heatmap
}
\examples{
with_options(bitmapType='cairo') ## disable plotting to X11
n <- 100
scale_mat <- matrix(rnorm(n**2, mean = 5), nrow = n)
rownames(scale_mat) <- paste0("G", 1:n)
colnames(scale_mat) <- paste0("S", 1:n)
top_genes <- paste0("G", n - 50:n - 20)
res <- single_kmeans_heatmap(scale_mat, 2, top_genes, "test", "test",
                            out_dir = tempdir(), "test", 7, 7)
}
