% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{top_n_genes}
\alias{top_n_genes}
\title{Top N Genes}
\usage{
top_n_genes(dsqres, top_ng, out_dir, prefix)
}
\arguments{
\item{dsqres}{the output of the DESeq2 function \code{results},
usually called after performing a contrast.}

\item{top_ng}{A positive integer for the N value of genes to
extract.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}

\item{prefix}{String to act as the prefix filename.}
}
\value{
A vector of N gene names
}
\description{
From a DESeq2 Result extract the top N genes as
determined by descending order of the minus Log10 Adjusted
P-value. Write the result to file as well as return it.
}
\examples{
dsqres <- data.frame(mLog10Padj = 1:10 / 10, gene = paste0("G", 1:10))
RNASeqHelper:::top_n_genes(dsqres, 3, out_dir = tempdir(), prefix = "test")
}
