% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Produce a Volcano Plot}
\usage{
volcano_plot(
    dsqres,
    degenes,
    title,
    curve = list(sd = 0.15, sc = 10, offset = 1),
    curve_scale = 1,
    curve_show = FALSE,
    ylim = NULL
)
}
\arguments{
\item{dsqres}{A table of DESeq2 results}

\item{degenes}{A vector of genes.}

\item{title}{A string depicting the title of the plot}

\item{curve}{A list of three components: sd, sc, offset.}

\item{curve_scale}{A numeric curve scaling factor.}

\item{curve_show}{A boolean on whether or not to show the curve.}

\item{ylim}{A vector of two components to limit the Y-axis.}
}
\value{
A ggplot2 object of the volcano plot.
}
\description{
Generates a volcano plot from the contrast of a DESeq2
analysis with added curves for graphic effect.
}
\examples{
n <- 100
dsqres <- data.frame(gene = paste0("G", 1:n),
                    data = rnorm(n, 5, 2) + 0.1,
                    log2FoldChange = rnorm(n, 0, 2) + 0.1,
                    padj = 0.1 + 1 / rnorm(n, 5, 2))
degenes <- paste0("G", (n-50):(n-20))
res <- volcano_plot(dsqres, degenes, "my plot")
}
