% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{do_gene_plots}
\alias{do_gene_plots}
\title{Perform Gene Plots}
\usage{
do_gene_plots(
    norm_counts,
    scale_mat,
    pheno_data,
    gene_cluster_scores,
    score_thresh,
    genes_of_interest,
    out_dir
)
}
\arguments{
\item{norm_counts}{a matrix of normalized count data, with genes as
rows and samples as columns.}

\item{scale_mat}{a matrix of scaled count data, with genes as rows
and samples as columns.}

\item{pheno_data}{a data.frame of phenotype data.}

\item{gene_cluster_scores}{A data frame with three columns: 'gene',
'cluster', and 'scores'.}

\item{score_thresh}{Vector of numerics depicting cluster score
thresholds to filter for high quality genes in each cluster
before plotting. Default is \code{c(0, 0.5, 0.9, 0.99)}}

\item{genes_of_interest}{A named list of gene vectors. A list of
genes which are referenced by specific names. Unique plots will
be generated for list. e.g. list(mygeneset1 = c("Msgn1",
"Osr1", "Rspo3", "Fgf8", "Wnt3a"), mygeneset2 = c("Mesp1",
"Foxa2", "Sox17", "Lhx1", "Cer1"))}

\item{out_dir}{A string denoting the output directory to store
plots and tables.}
}
\value{
None.
}
\description{
Perform cluster gene plots for normalised and scaled
data as well as plotting genes of interest for both.
}
\examples{
n <- 100
norm_counts <- matrix(rnorm(n**2, mean = 5), nrow = n)
rownames(norm_counts) <- paste0("G", 1:n)
colnames(norm_counts) <- paste0("S", 1:n)
scale_mat <- matrix(rnorm(n**2, mean = 0), nrow = n)
rownames(scale_mat) <- paste0("G", 1:n)
colnames(scale_mat) <- paste0("S", 1:n)
pheno <- data.frame(
    sample = paste0("S", 1:n),
    condition = c(rep("red", n / 2), rep("green", n / 2)),
    time = as.integer(rnorm(n, 2, 0.5) + 1) * 5
)
gene_cluster_scores <- data.frame(
    gene = rownames(scale_mat),
    cluster = c(rep(1, n / 50), rep(2, n / 50)),
    score = rnorm(n, 0.5, 0.25)
)
score_thresh <- c(0.2, 0.3)
genes_of_interest <- paste0("G", 5:50)
res <- do_gene_plots(norm_counts, scale_mat, pheno,
                    gene_cluster_scores, score_thresh,
                    genes_of_interest,
                    out_dir = "/tmp")
}
