% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{pairwise_hmap_volcano}
\alias{pairwise_hmap_volcano}
\title{Pairwise heatmaps and volcano plots}
\usage{
pairwise_hmap_volcano(
    ddsObj,
    transformed_counts = NULL,
    numer,
    denom,
    top_ngenes_tocluster = 2000,
    top_ngenes_tohighlight = 50,
    score_thresh = c(0.5, 0.9),
    genes_of_interest = NULL,
    volcano_params = NULL,
    kmeans_list = c(2, 5, 8, 16),
    out_dirprefix = "."
)
}
\arguments{
\item{ddsObj}{A DESeq object,}

\item{transformed_counts}{REALLY UNSURE WHAT'S GOING ON HERE}

\item{numer}{A string value to select sample columns from the
\code{condition} column that will be part of the numerator part
of the contrast.}

\item{denom}{A string value to select sample columns from the
\code{condition} column, that will be the denominator part of
the contrast.}

\item{top_ngenes_tocluster}{A positive integer. How many of the top
log10 padj values to take. Default is 2000.}

\item{top_ngenes_tohighlight}{A positive integer. How many of the
top log10 padj values to label in the plots. Default is 50.}

\item{score_thresh}{Vector of numerics depicting cluster score
thresholds to filter for high quality genes in each cluster
before plotting. Default is \code{c(0, 0.5, 0.9, 0.99)}}

\item{genes_of_interest}{A named list of gene groups to plot in
ribbot plots. If NULL}

\item{volcano_params}{A list of two components: "global" and
"zoomed" each with "curve" and "curve_show" components
describing the curve to fit to the volcano plots by \code{sd}
(standard deviation from center mean), \code{sc} vertical scale
factor, \code{offset} vertical offset. Extra params for the
\code{zoomed} component are \code{ylim}, \code{padj} an upper
limit on the adjusted p-value, \code{lfc} and an upper-limit on
the log2 fold change. If \code{curve_show} is set to FALSE in
either \code{global} or \code{zoomed} then no curve is shown.}

\item{kmeans_list}{A list of positive integers to do k-means}

\item{out_dirprefix}{A character prefix outlining the directory and
basename in which plots and tables will be deployed.}
}
\value{
Void. Plots are deposited to the output directory.
}
\description{
For a given contrast and DESeq object produce volcano
plots of the DE genes, and generate a global heatmap with those
DE genes and a pairwise one with just the numerator and
denominator samples.
}
