% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqhelper.R
\name{run_deseq}
\alias{run_deseq}
\title{Run DESeq with sensible defaults}
\usage{
run_deseq(tab, keep_genes, phenotype_data, out_dir)
}
\arguments{
\item{tab}{a matrix of samples (columns) and informative genes
(rows), ideally subset using the function
\code{high_quality_genes}}

\item{keep_genes}{a vector of genes to subset.}

\item{phenotype_data}{a table with samples (rows) with extra
columns for annotation groups. One of these groups must be
condition and will be used to normalize the data.}

\item{out_dir}{String depicting the output directory to place
tables and plots.}
}
\value{
list of tables
}
\description{
Runs DESeq with filtering thresholds and shows PCA for
variance-stabilize-transformed data.
}
\examples{
n <- 1000
tab <- matrix(as.integer(rnorm(n**2, 1000, 500)), ncol = n)
tab <- tab - min(tab)
rownames(tab) <- paste0("G", 1:n)
colnames(tab) <- paste0("S", 1:n)
phenotype_data <- data.frame(
    sample = paste0("S", 1:n),
    condition = c(rep("red", n / 2), rep("green", n / 2)),
    time = as.integer(rnorm(n, 2, 0.5) + 1) * 5
)
keep_genes <- paste0("G", 1:n)
res <- run_deseq(tab, keep_genes, phenotype_data, tempdir())
}
